% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{countdown}
\alias{countdown}
\title{Countdown before deadline.}
\usage{
countdown(deadline, complete_str = intToUtf8(10004))
}
\arguments{
\item{deadline}{POSIXct. deadline}

\item{complete_str}{string. displayed when deadline is passed}
}
\description{
Countdown before deadline.
}

