% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\name{data_split}
\alias{data_split}
\title{Split a data.frame into training and test sets.}
\usage{
data_split(data = get_data("german"), varname = "Class", p_test = 0.2,
  p_quiz = 0.5)
}
\arguments{
\item{data}{data.frame}

\item{varname}{string. output variable name}

\item{p_test}{real. proportion of samples in the test set}

\item{p_quiz}{real. proportion of samples from the test set in the quiz set}
}
\value{
list with members
  \item{train}{training set with output variable}
  \item{test}{test set without output variable}
  \item{y_test}{test set output variable}
  \item{ind_quiz}{indices of quiz samples in the test set}
}
\description{
Split a data.frame into training and test sets.
}

