% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{fa}
\alias{fa}
\title{HTML code for a Font Awesome icon.}
\usage{
fa(name)
}
\arguments{
\item{name}{string. name of the icon.}
}
\value{
string containing the HTML code.
}
\description{
HTML code for a Font Awesome icon.
}
\note{
Requires the Font Awesome HTML code:
  \code{<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css">}
}
\seealso{
\url{http://fontawesome.io/}
}

