% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{german}
\alias{german}
\title{German Credit Data}
\format{A \code{data.frame} with 1000 rows and 21 variables}
\source{
UCI Machine Learning Repository
  \url{https://archive.ics.uci.edu/ml/datasets/Statlog+(German+Credit+Data)}
}
\usage{
german
}
\description{
Data from Dr. Hans Hofmann of the University of Hamburg.
}
\details{
These data have two classes for the credit worthiness: Good or Bad. There are
predictors related to attributes, such as: checking account status, duration,
credit history, purpose of the loan, amount of the loan, savings accounts or
bonds, employment duration, Installment rate in percentage of disposable income,
personal information, other debtors/guarantors, residence duration, property, age,
other installment plans, housing, number of existing credits, job information,
Number of people being liable to provide maintenance for, telephone, and foreign
worker status.

This is a transformed version of the \code{\link[caret]{GermanCredit}} data set
with factors instead of dummy variables
}
\keyword{datasets}

