% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\name{get_data}
\alias{get_data}
\title{Get dataset value.}
\usage{
get_data(name = "german", package = "rchallenge", envir = environment(),
  ...)
}
\arguments{
\item{name}{string. name of the dataset.}

\item{package}{string. name of the package to look in for dataset.}

\item{envir}{the environment where the data should be loaded.}

\item{...}{additional arguments to be passed to \code{\link[utils]{data}}.}
}
\value{
The value of the dataset
}
\description{
Get dataset value.
}
\seealso{
\code{\link[utils]{data}}, \code{\link{base}}
}

