% Generated by roxygen2 (4.0.2): do not edit by hand
\name{last_update}
\alias{last_update}
\title{Formatted last date before deadline}
\usage{
last_update(deadline, format = "\%d \%b \%Y \%H:\%M")
}
\arguments{
\item{deadline}{POSIXct. deadline}

\item{format}{string. see \code{\link{format.POSIXct}}}
}
\description{
Formatted last date before deadline
}

