% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_challenges}
\alias{list_challenges}
\title{List all available challenge ids}
\usage{
list_challenges()
}
\value{
A character vector of challenge ids.
}
\description{
Returns the ids of every challenge currently registered in the
package. Use an id with \code{\link[=run_challenge]{run_challenge()}} to attempt that challenge.
}
\examples{
list_challenges()

}
