% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print_readerr}
\alias{print_readerr}
\title{Print read errors}
\usage{
print_readerr(read_err = list())
}
\arguments{
\item{read_err}{list of read errors returned by \code{\link{store_new_submissions}}}
}
\value{
\code{NULL}
}
\description{
Print read errors
}

