% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{rc_longest_streak}
\alias{rc_longest_streak}
\title{Calculate the longest streak ever achieved (not just the current one)}
\usage{
rc_longest_streak()
}
\description{
Calculate the longest streak ever achieved (not just the current one)
}
\keyword{internal}
