% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchallenge.r
\name{rchallenge-defunct}
\alias{glyphicon}
\alias{rchallenge-defunct}
\title{Defunct functions in package \sQuote{rchallenge}}
\usage{
glyphicon(...)
}
\arguments{
\item{...}{parameters}
}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: \code{glyphicon}
}

