% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_challenge}
\alias{register_challenge}
\title{Register a challenge into the internal registry}
\usage{
register_challenge(challenge)
}
\description{
Register a challenge into the internal registry
}
\keyword{internal}
