% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display.r
\name{str_rank}
\alias{str_rank}
\title{String displayed for the rank.}
\usage{
str_rank(r, r_d, symb = list(const = html_img(glyphicon("right_arrow"),
  "10px"), up = html_img(glyphicon("up_arrow"), "10px"), down =
  html_img(glyphicon("down_arrow"), "10px")))
}
\arguments{
\item{r}{integer. rank}

\item{r_d}{integer. rank difference}

\item{symb}{named list of characters. symbols used for the progress in ranking:
no change (\code{const}), ascent (\code{up}) and descent (\code{down})}
}
\description{
Concatenates the rank number with symbols indicating the progress since the last change.
}
\keyword{internal}

