% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.r
\name{countdown}
\alias{countdown}
\title{Countdown before deadline}
\usage{
countdown(dl = deadline, complete_str = intToUtf8(10004))
}
\arguments{
\item{dl}{POSIXct. deadline}

\item{complete_str}{string. displayed when deadline is passed}
}
\description{
Countdown before deadline
}

