% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.r
\name{glyphicon}
\alias{glyphicon}
\title{Path to glyphicon image file}
\usage{
glyphicon(name, path = system.file("glyphicons", package = "challenge"))
}
\arguments{
\item{name}{string. name of the glyphicon.}

\item{path}{string. folder of search.}
}
\value{
the path to the file.
}
\description{
Path to glyphicon image file
}

