% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{html_img}
\alias{html_img}
\title{HTML code for an image.}
\usage{
html_img(file, width = "10px")
}
\arguments{
\item{file}{string. image file.}

\item{width}{string. width of display.}
}
\description{
HTML code for an image.
}
