% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.r
\name{last_update}
\alias{last_update}
\title{Formatted last date before deadline}
\usage{
last_update(dl = deadline + 1, format = "\%d \%b \%Y \%H:\%M")
}
\arguments{
\item{dl}{POSIXct. deadline}

\item{format}{string. see \code{\link{format.POSIXct}}}
}
\description{
Formatted last date before deadline
}

