% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{last_update}
\alias{last_update}
\title{Formatted last update date before deadline.}
\usage{
last_update(deadline, format = "\%d \%b \%Y \%H:\%M")
}
\arguments{
\item{deadline}{POSIXct. deadline}

\item{format}{string. see \code{\link{format.POSIXct}}}
}
\description{
Formatted last update date before deadline.
}

