% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new_team}
\alias{new_team}
\title{Create a new team submissions folder in your challenge}
\usage{
new_team(name, path = ".", submissions_dir = "submissions", quiet = FALSE,
  showWarnings = FALSE)
}
\arguments{
\item{name}{string. name of the team subdirectory}

\item{path}{string. root path of the challenge. see \code{\link{new_challenge}}}

\item{submissions_dir}{string. subdirectory of the submissions. see \code{\link{new_challenge}}}

\item{quiet}{logical. deactivate text output.}

\item{showWarnings}{logical. should the warnings on failure be shown? see \code{\link{dir.create}}}
}
\description{
Create a new team submissions folder in your challenge
}

