% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{print_leaderboard}
\alias{print_leaderboard}
\title{Format the leaderboard in Markdown.}
\usage{
print_leaderboard(best, metric, test_name = "quiz", ...)
}
\arguments{
\item{best}{list of the best submissions per team and per metric as returned
by \code{\link{get_best}}.}

\item{metric}{string. name of the metric considered}

\item{test_name}{string. name of the test set used: \code{"quiz"} or \code{"test"}}

\item{...}{further parameters to pass to \code{\link[knitr]{kable}}}
}
\value{
\code{print_leaderboard} returns a character vector of the table source code
  to be used in a Markdown document.
}
\description{
Format the leaderboard in Markdown.
}
\note{
Chunk option \code{results='asis'} has to be used
}
\seealso{
\code{\link[knitr]{kable}}
}

