% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{str_rank}
\alias{str_rank}
\title{String displayed for the rank.}
\usage{
str_rank(r, r_d, symb = list(const = icon("fa-arrow-right"), up =
  icon("fa-arrow-up"), down = icon("fa-arrow-down")))
}
\arguments{
\item{r}{integer. rank}

\item{r_d}{integer. rank difference}

\item{symb}{named list of characters. symbols used for the progress in ranking:
no change (\code{const}), ascent (\code{up}) and descent (\code{down})}
}
\description{
Concatenates the rank number with symbols indicating the progress since the last change.
}
\keyword{internal}

